chore: split out utils, continue cleanup
