+++ src/components/command-bar/commands/search/unified-search/components/dialog-body/dialog-body.module.css
@@ -1,0 +1,3 @@
+.suggestedPagesWrapper {
+	padding: 16px;
+}
