--- src/components/command-bar/commands/search/unified-search/components/unified-hit/render-highlight-array.ts
@@ -1,14 +1,0 @@
-export function renderHighlightArray(facet, matchesOnly = false) {
-	return (facet || [])
-		.filter((entry) => {
-			if (matchesOnly) {
-				return entry.matchLevel !== 'none'
-			} else {
-				return true
-			}
-		})
-		.map((entry) => {
-			return entry?.value
-		})
-		.join(', ')
-}
