+++ src/components/command-bar/commands/search/unified-search/utils/build-algolia-filters.ts
