+++ src/components/command-bar/commands/search/unified-search/utils/build-url-path.ts
