+++ src/components/command-bar/commands/search/unified-search/utils/get-algolia-content-type.ts
