+++ src/components/command-bar/commands/search/unified-search/utils/get-should-render-integrations-tab.ts
